refactor CalendarScreen and DayInCalendarScreen
in addition to the location and other minor changes, some repeated
logic for the logs creation and prediction have been centralized.
