+++ src/screens/DayInCalendarScreen.js
@@ -3,13 +3,12 @@ import { View, Pressable } from 'react-native'
 import { connect } from 'react-redux';
 import { Text, Subheading } from 'react-native-paper'
 import { FontAwesomeIcon } from '@fortawesome/react-native-fontawesome'
-import { faCalendarDay, faCalendarWeek, faSlash } from '@fortawesome/free-solid-svg-icons'
+import { faCalendarDay, faCalendarWeek } from '@fortawesome/free-solid-svg-icons'
 import { DateTime } from 'luxon'
-import { selectEntriesByDay, selectActivityById, selectActivityEntities, selectGoalById } from '../../redux'
-import { Header, ActivityList } from '../../components' 
-import { extractActivityLists, dueToday, newEntry, getToday } from '../../util'
 import { useTranslation } from 'react-i18next'
-import { CalendarColor, GeneralColor } from '../../styles/Colors';
+import { Header, ActivityList } from '../components' 
+import { extractActivityLists } from '../util'
+import { CalendarColor, GeneralColor } from '../styles/Colors';
 
 const DayInCalendarScreen = ({ navigation, dayActivities, weekActivities, day }) => {
   const [ viewMode, setViewMode ] = React.useState('day') // can be 'day' or 'week'
@@ -38,38 +37,14 @@ const DayInCalendarScreen = ({ navigation, dayActivities, weekActivities, day })
   )
 }
 
-function predictEntries(state, day){
-  let entries = []
-
-  const activities = selectActivityEntities(state)
-  for(let activityId in activities){
-    const activity = activities[activityId]
-    const goal = selectGoalById(state, activity.goalId)
-
-    if(dueToday(day, activity, goal)){
-      entries.push(newEntry(activity))
-    }
-  }
-
-  return entries
-}
-
 const mapStateToProps = (state, ownProps) => {
   const dateString = ownProps.route.params.day
   const day = DateTime.fromFormat(dateString, 'yyyy-MM-dd')
 
-  let dayActivities, weekActivities
-  if(day > getToday(state.settings.dayStartHour)){
-    const predictedEntries = predictEntries(state, day);
-    // automatic semicolon insertion fails when line starts with "opening something"...
-    ;({ dayActivities, weekActivities } = extractActivityLists(state, day, predictedEntries));
-  }else{
-    ;({ dayActivities, weekActivities } = extractActivityLists(state, day));
-  }
+  const { dayActivities, weekActivities } = extractActivityLists(state, day)
 
   return { dayActivities, weekActivities, day }
 }
-
 
 const actionToProps = {
 }
