+++ src/screens/index.js
@@ -5,7 +5,7 @@ export { default as GoalsScreen } from './GoalsScreen'
 export { default as GoalScreen} from './GoalScreen'
 export { default as ActivityFormScreen } from './ActivityFormScreen/ActivityFormScreen'
 export { default as GoalFormScreen } from './GoalFormScreen'
-export { default as CalendarScreen } from './CalendarScreen/CalendarScreen'
+export { default as CalendarScreen } from './CalendarScreen'
 export { default as SettingsScreen } from './SettingsScreen'
-export { default as DayInCalendarScreen } from './DayInCalendarScreen/DayInCalendarScreen'
+export { default as DayInCalendarScreen } from './DayInCalendarScreen'
 export { default as OnboardingScreen } from './OnboardingScreen'
